Add files via upload
Submission, upto 382512pm, make a copy in this folder and use for edits, inside Visual Studio and as each feature is added, give a description, with each, edited file update in here, to help map progress and milestones, don't forget date and time, for each file. Perhaps have, Major and Minor Files. Whatever, creates a "Hierarchy", of the codes evolution today 3825.